Compress color runs, add error handling
Also add a better test file.

___________________________

Runs of same-color pixels are represented by a single rect element.

If the image-parsing method fails, a message with a link to the file is
printed to the page.

The order of attributes has been shuffled to maximize the length of
constant repeated strings.

The unused code checking file extension has been removed.