Set MIME type, make responsive, include sample output
The code now outputs a complete SVG file if there is no error, including
MIME type header and SVG namespace, so the resulting file can be saved
as a valid .svg.

It also adds a viewBox attribute so the image becomes responsive.  When
displayed full screen, the results nicely demonstrate browsers' horrible
rendering rounding errors in SVG, so a `shape-rendering: crispEdges`
setting is added, which vastly improves the appearance in Firefox and
Chrome.

The sample files are the result of running the PHP on the [new test
image][1], a 22.4 KB PNG.  This creates a 489 KB raw SVG file, or 43 KB
after gzip compression with 7zip.

[1]:
https://pbs.twimg.com/profile_images/539094743133614080/cj8fsBZE_reasonably_small.png